Refactor boundary handler out of main function
