+++ pkg/serverus/types.go
@@ -1,0 +1,53 @@
+package serverus
+
+/*
+	type declaration file used to hold types shared between the
+	main package and serverus
+*/
+
+type Game struct {
+	ID      string `json:"id"`
+	Timeout int32  `json:"timeout"`
+}
+
+type Coord struct {
+	X int `json:"x"`
+	Y int `json:"y"`
+}
+
+type Battlesnake struct {
+	ID     string  `json:"id"`
+	Name   string  `json:"name"`
+	Health int32   `json:"health"`
+	Body   []Coord `json:"body"`
+	Head   Coord   `json:"head"`
+	Length int32   `json:"length"`
+	Shout  string  `json:"shout"`
+}
+
+type Board struct {
+	Height int           `json:"height"`
+	Width  int           `json:"width"`
+	Food   []Coord       `json:"food"`
+	Snakes []Battlesnake `json:"snakes"`
+}
+
+type BattlesnakeInfoResponse struct {
+	APIVersion string `json:"apiversion"`
+	Author     string `json:"author"`
+	Color      string `json:"color"`
+	Head       string `json:"head"`
+	Tail       string `json:"tail"`
+}
+
+type GameRequest struct {
+	Game  Game        `json:"game"`
+	Turn  int         `json:"turn"`
+	Board Board       `json:"board"`
+	You   Battlesnake `json:"you"`
+}
+
+type MoveResponse struct {
+	Move  string `json:"move"`
+	Shout string `json:"shout,omitempty"`
+}
